+++ src/plasp/sas/TranslatorASP.cpp
@@ -1,5 +1,7 @@
 #include <plasp/sas/TranslatorASP.h>
 
+#include <plasp/utils/Formatting.h>
+
 namespace plasp
 {
 namespace sas
@@ -11,9 +13,9 @@ namespace sas
 //
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
-TranslatorASP::TranslatorASP(const Description &description, std::ostream &ostream)
+TranslatorASP::TranslatorASP(const Description &description, utils::LogStream &ostream)
 :	m_description(description),
-	m_ostream(ostream)
+	m_outputStream(ostream)
 {
 }
 
@@ -34,67 +36,67 @@ void TranslatorASP::translate() const
 
 void TranslatorASP::translateRequirements() const
 {
-	m_ostream << "% feature requirements" << std::endl;
+	m_outputStream << utils::Heading2("feature requirements") << std::endl;
 
 	if (m_description.usesActionCosts())
-		m_ostream << "requiresFeature(actionCosts)." << std::endl;
+		m_outputStream << utils::Keyword("requiresFeature") << "(actionCosts)." << std::endl;
 
 	if (m_description.usesAxiomRules())
-		m_ostream << "requiresFeature(axiomRules)." << std::endl;
+		m_outputStream << utils::Keyword("requiresFeature") << "(axiomRules)." << std::endl;
 
 	if (m_description.usesConditionalEffects())
-		m_ostream << "requiresFeature(conditionalEffects)." << std::endl;
+		m_outputStream << utils::Keyword("requiresFeature") << "(conditionalEffects)." << std::endl;
 
-	m_ostream << std::endl;
+	m_outputStream << std::endl;
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
 void TranslatorASP::translateInitialState() const
 {
-	m_ostream << "% initial state" << std::endl;
+	m_outputStream << utils::Heading2("initial state") << std::endl;
 
 	const auto &initialStateFacts = m_description.initialState().facts();
 
 	std::for_each(initialStateFacts.cbegin(), initialStateFacts.cend(),
 		[&](const auto &fact)
 		{
-			m_ostream << "initialState(";
-			fact.variable().printNameAsASPPredicate(m_ostream);
-			m_ostream << ", ";
-			fact.value().printAsASPPredicate(m_ostream);
-			m_ostream << ")." << std::endl;
+			m_outputStream << utils::Keyword("initialState") << "(";
+			fact.variable().printNameAsASPPredicate(m_outputStream);
+			m_outputStream << ", ";
+			fact.value().printAsASPPredicate(m_outputStream);
+			m_outputStream << ")." << std::endl;
 		});
 
-	m_ostream << std::endl;
+	m_outputStream << std::endl;
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
 void TranslatorASP::translateGoal() const
 {
-	m_ostream << "% goal" << std::endl;
+	m_outputStream << utils::Heading2("goal") << std::endl;
 
 	const auto &goalFacts = m_description.goal().facts();
 
 	std::for_each(goalFacts.cbegin(), goalFacts.cend(),
 		[&](const auto &fact)
 		{
-			m_ostream << "goal(";
-			fact.variable().printNameAsASPPredicate(m_ostream);
-			m_ostream << ", ";
-			fact.value().printAsASPPredicate(m_ostream);
-			m_ostream << ")." << std::endl;
+			m_outputStream << utils::Keyword("goal") << "(";
+			fact.variable().printNameAsASPPredicate(m_outputStream);
+			m_outputStream << ", ";
+			fact.value().printAsASPPredicate(m_outputStream);
+			m_outputStream << ")." << std::endl;
 		});
 
-	m_ostream << std::endl;
+	m_outputStream << std::endl;
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
 void TranslatorASP::translateVariables() const
 {
-	m_ostream << "% variables";
+	m_outputStream << utils::Heading2("variables");
 
 	const auto &variables = m_description.variables();
 
@@ -105,51 +107,51 @@ void TranslatorASP::translateVariables() const
 
 			BOOST_ASSERT(!values.empty());
 
-			m_ostream << std::endl;
-			variable.printNameAsASPPredicate(m_ostream);
-			m_ostream << "." << std::endl;
+			m_outputStream << std::endl;
+			variable.printNameAsASPPredicate(m_outputStream);
+			m_outputStream << "." << std::endl;
 
 			std::for_each(values.cbegin(), values.cend(),
 				[&](const auto &value)
 				{
-					m_ostream << "contains(";
-					variable.printNameAsASPPredicate(m_ostream);
-					m_ostream << ", ";
-					value.printAsASPPredicate(m_ostream);
-					m_ostream << ")." << std::endl;
+					m_outputStream << utils::Keyword("contains") << "(";
+					variable.printNameAsASPPredicate(m_outputStream);
+					m_outputStream << ", ";
+					value.printAsASPPredicate(m_outputStream);
+					m_outputStream << ")." << std::endl;
 				});
 		});
 
-	m_ostream << std::endl;
+	m_outputStream << std::endl;
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
 void TranslatorASP::translateActions() const
 {
-	m_ostream << "% actions";
+	m_outputStream << utils::Heading2("actions");
 
 	const auto &operators = m_description.operators();
 
 	std::for_each(operators.cbegin(), operators.cend(),
 		[&](const auto &operator_)
 		{
-			m_ostream << std::endl;
-			operator_.printPredicateAsASP(m_ostream);
-			m_ostream << "." << std::endl;
+			m_outputStream << std::endl;
+			operator_.printPredicateAsASP(m_outputStream);
+			m_outputStream << "." << std::endl;
 
 			const auto &preconditions = operator_.preconditions();
 
 			std::for_each(preconditions.cbegin(), preconditions.cend(),
 				[&](const auto &precondition)
 				{
-					m_ostream << "precondition(";
-					operator_.printPredicateAsASP(m_ostream);
-					m_ostream << ", ";
-					precondition.variable().printNameAsASPPredicate(m_ostream);
-					m_ostream << ", ";
-					precondition.value().printAsASPPredicate(m_ostream);
-					m_ostream << ")." << std::endl;
+					m_outputStream << utils::Keyword("precondition") << "(";
+					operator_.printPredicateAsASP(m_outputStream);
+					m_outputStream << ", ";
+					precondition.variable().printNameAsASPPredicate(m_outputStream);
+					m_outputStream << ", ";
+					precondition.value().printAsASPPredicate(m_outputStream);
+					m_outputStream << ")." << std::endl;
 				});
 
 			const auto &effects = operator_.effects();
@@ -164,39 +166,39 @@ void TranslatorASP::translateActions() const
 					std::for_each(conditions.cbegin(), conditions.cend(),
 						[&](const auto &condition)
 						{
-							m_ostream << "effectCondition(";
-							operator_.printPredicateAsASP(m_ostream);
-							m_ostream << ", effect(" << currentEffectID << "), ";
-							condition.variable().printNameAsASPPredicate(m_ostream);
-							m_ostream << ", ";
-							condition.value().printAsASPPredicate(m_ostream);
-							m_ostream << ")." << std::endl;
+							m_outputStream << utils::Keyword("effectCondition") << "(";
+							operator_.printPredicateAsASP(m_outputStream);
+							m_outputStream << ", " << utils::Keyword("effect") << "(" << currentEffectID << "), ";
+							condition.variable().printNameAsASPPredicate(m_outputStream);
+							m_outputStream << ", ";
+							condition.value().printAsASPPredicate(m_outputStream);
+							m_outputStream << ")." << std::endl;
 						});
 
-					m_ostream << "postcondition(";
-					operator_.printPredicateAsASP(m_ostream);
-					m_ostream << ", effect(" << currentEffectID << "), ";
-					effect.postcondition().variable().printNameAsASPPredicate(m_ostream);
-					m_ostream << ", ";
-					effect.postcondition().value().printAsASPPredicate(m_ostream);
-					m_ostream << ")." << std::endl;
+					m_outputStream << utils::Keyword("postcondition") << "(";
+					operator_.printPredicateAsASP(m_outputStream);
+					m_outputStream << ", " << utils::Keyword("effect") << "(" << currentEffectID << "), ";
+					effect.postcondition().variable().printNameAsASPPredicate(m_outputStream);
+					m_outputStream << ", ";
+					effect.postcondition().value().printAsASPPredicate(m_outputStream);
+					m_outputStream << ")." << std::endl;
 
 					currentEffectID++;
 				});
 
-			m_ostream << "costs(";
-			operator_.printPredicateAsASP(m_ostream);
-			m_ostream << ", " << operator_.costs() << ")." << std::endl;
+			m_outputStream << utils::Keyword("costs") << "(";
+			operator_.printPredicateAsASP(m_outputStream);
+			m_outputStream << ", " << operator_.costs() << ")." << std::endl;
 		});
 
-	m_ostream << std::endl;
+	m_outputStream << std::endl;
 }
 
 ////////////////////////////////////////////////////////////////////////////////////////////////////
 
 void TranslatorASP::translateMutexes() const
 {
-	m_ostream << "% mutex groups";
+	m_outputStream << utils::Heading2("mutex groups");
 
 	const auto &mutexGroups = m_description.mutexGroups();
 
@@ -208,18 +210,18 @@ void TranslatorASP::translateMutexes() const
 			const auto mutexGroupID = std::to_string(currentMutexGroupID);
 			currentMutexGroupID++;
 
-			m_ostream << std::endl << "mutexGroup(" << mutexGroupID << ")." << std::endl;
+			m_outputStream << std::endl << utils::Keyword("mutexGroup") << "(" << mutexGroupID << ")." << std::endl;
 
 			const auto &facts = mutexGroup.facts();
 
 			std::for_each(facts.cbegin(), facts.cend(),
 				[&](const auto &fact)
 				{
-					m_ostream << "contains(mutexGroup(" << mutexGroupID << "), ";
-					fact.variable().printNameAsASPPredicate(m_ostream);
-					m_ostream << ", ";
-					fact.value().printAsASPPredicate(m_ostream);
-					m_ostream << ")." << std::endl;
+					m_outputStream << utils::Keyword("contains") << "(" << utils::Keyword("mutexGroup") << "(" << mutexGroupID << "), ";
+					fact.variable().printNameAsASPPredicate(m_outputStream);
+					m_outputStream << ", ";
+					fact.value().printAsASPPredicate(m_outputStream);
+					m_outputStream << ")." << std::endl;
 				});
 		});
 }
@@ -231,8 +233,8 @@ void TranslatorASP::translateAxiomRules() const
 	if (!m_description.usesActionCosts())
 		return;
 
-	m_ostream << std::endl;
-	m_ostream << "% axiom rules";
+	m_outputStream << std::endl;
+	m_outputStream << utils::Heading2("axiom rules");
 
 	const auto &axiomRules = m_description.axiomRules();
 
@@ -244,27 +246,27 @@ void TranslatorASP::translateAxiomRules() const
 			const auto axiomRuleID = std::to_string(currentAxiomRuleID);
 			currentAxiomRuleID++;
 
-			m_ostream << std::endl << "axiomRule(" << axiomRuleID << ")." << std::endl;
+			m_outputStream << std::endl << utils::Keyword("axiomRule") << "(" << axiomRuleID << ")." << std::endl;
 
 			const auto &conditions = axiomRule.conditions();
 
 			std::for_each(conditions.cbegin(), conditions.cend(),
 				[&](const auto &condition)
 				{
-					m_ostream << "condition(axiomRule(" << axiomRuleID << "), ";
-					condition.variable().printNameAsASPPredicate(m_ostream);
-					m_ostream << ", ";
-					condition.value().printAsASPPredicate(m_ostream);
-					m_ostream << ")." << std::endl;
+					m_outputStream << utils::Keyword("condition") << "(" << utils::Keyword("axiomRule") << "(" << axiomRuleID << "), ";
+					condition.variable().printNameAsASPPredicate(m_outputStream);
+					m_outputStream << ", ";
+					condition.value().printAsASPPredicate(m_outputStream);
+					m_outputStream << ")." << std::endl;
 				});
 
 			const auto &postcondition = axiomRule.postcondition();
 
-			m_ostream << "postcondition(axiomRule(axiomRule" << axiomRuleID << "), ";
-			postcondition.variable().printNameAsASPPredicate(m_ostream);
-			m_ostream << ", ";
-			postcondition.value().printAsASPPredicate(m_ostream);
-			m_ostream << ")." << std::endl;
+			m_outputStream << utils::Keyword("postcondition") << "(" << utils::Keyword("axiomRule") << "(" << axiomRuleID << "), ";
+			postcondition.variable().printNameAsASPPredicate(m_outputStream);
+			m_outputStream << ", ";
+			postcondition.value().printAsASPPredicate(m_outputStream);
+			m_outputStream << ")." << std::endl;
 		});
 }